pyproject.toml
[tool.poetry]
name = "carreer-finder"
version = "0.1.0"
description = ""
authors = ["Zach Smith <zach.smith2171@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
supabase = "^2.7.4"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
